requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
alembic==1.16.4
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic[email]==2.11.7
email-validator==2.2.0 
bcrypt==4.3.0
python-jose==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
isort==6.0.1
black==25.1.0
flake8==7.3.0
pre-commit==4.3.0
psycopg2-binary==2.9.10
fastapi[standard]==0.116.1